Move content_item_selection.py to legacy_content_item_selection.py
