+++ lms/views/legacy_content_item_selection.py
