+++ lms/views/lti_launches.py
@@ -8,7 +8,7 @@ from lms.models.lti_launches import LtiLaunches
 from lms.models.module_item_configuration import ModuleItemConfiguration
 from lms.util import lti_launch
 from lms.util.view_renderer import view_renderer
-from lms.views.content_item_selection import content_item_form
+from lms.views.legacy_content_item_selection import content_item_form
 from lms.util.associate_user import associate_user
 from lms.util.canvas_api import CanvasApi, GET
 from lms.util.authorize_lms import authorize_lms, save_token
